Implemented custom actions on vendor overview page
