+++ src/pages/purchase-orders/common/hooks/useCustomBulkActions.tsx
@@ -1,0 +1,36 @@
+/**
+ * Invoice Ninja (https://invoiceninja.com).
+ *
+ * @link https://github.com/invoiceninja/invoiceninja source repository
+ *
+ * @copyright Copyright (c) 2022. Invoice Ninja LLC (https://invoiceninja.com)
+ *
+ * @license https://www.elastic.co/licensing/elastic-license
+ */
+
+import { PurchaseOrder } from '$app/common/interfaces/purchase-order';
+import { CustomBulkAction } from '$app/components/DataTable';
+import { DropdownElement } from '$app/components/dropdown/DropdownElement';
+import { Icon } from '$app/components/icons/Icon';
+import { usePrintPdf } from '$app/pages/invoices/common/hooks/usePrintPdf';
+import { useTranslation } from 'react-i18next';
+import { MdPrint } from 'react-icons/md';
+
+export function useCustomBulkActions() {
+  const [t] = useTranslation();
+
+  const printPdf = usePrintPdf({ entity: 'purchase_order' });
+
+  const customBulkActions: CustomBulkAction<PurchaseOrder>[] = [
+    (selectedIds) => (
+      <DropdownElement
+        onClick={() => printPdf(selectedIds)}
+        icon={<Icon element={MdPrint} />}
+      >
+        {t('print_pdf')}
+      </DropdownElement>
+    ),
+  ];
+
+  return customBulkActions;
+}
